+++ app/(tabs)/settings.tsx
@@ -1,6 +1,5 @@
 import ParallaxScrollView from "@/components/parallax-scroll-view";
 import ChangeThemeButton from "@/components/settings/changeThemeButton";
-import DevToolsButton from "@/components/settings/devToolsButton";
 import LogoutButton from "@/components/settings/logoutButton";
 import TaskFontSizeButton from "@/components/settings/taskFontSizeButton";
 import VersionInfoButton from "@/components/settings/versionInfoButton";
@@ -46,7 +45,7 @@ export default function SettingsPage() {
         <ChangeThemeButton />
         <TaskFontSizeButton />
         <VersionInfoButton />
-        <DevToolsButton />
+        {/* <DevToolsButton /> */}
         <LogoutButton />
       </ThemedView>
       <ThemedText style={dynamicStyles.text}>